feat: archived groups and memberships

+++ src/modules/user.methods.ts
@@ -1,4 +1,5 @@
 import { BonusAction, User } from 'prisma/prisma-client';
+import { TRPCError } from '@trpc/server';
 
 import { prisma } from '../utils/prisma';
 import { SessionUser } from '../utils/auth';
@@ -6,6 +7,7 @@ import { SessionUser } from '../utils/auth';
 import { MembershipInfo, UserMemberships, UserMeta, UserSupervisor } from './user.types';
 import { AddUserToGroup, ChangeBonusPoints, EditUser, GetUserList, RemoveUserFromGroup } from './user.schemas';
 import { userAccess } from './user.access';
+import { tr } from './modules.i18n';
 
 export const addCalculatedUserFields = <T extends User>(user: T, sessionUser: SessionUser): T & UserMeta => {
     return {
@@ -19,11 +21,19 @@ export const addCalculatedUserFields = <T extends User>(user: T, sessionUser: Se
 };
 
 export const userMethods = {
-    addToGroup: (data: AddUserToGroup) => {
+    addToGroup: async (data: AddUserToGroup) => {
+        const membership = await prisma.membership.findUnique({ where: { userId_groupId: data } });
+        if (membership?.archived) {
+            throw new TRPCError({ code: 'BAD_REQUEST', message: tr('Cannot edit archived membership') });
+        }
         return prisma.membership.create({ data });
     },
 
-    removeFromGroup: (data: RemoveUserFromGroup) => {
+    removeFromGroup: async (data: RemoveUserFromGroup) => {
+        const membership = await prisma.membership.findUnique({ where: { userId_groupId: data } });
+        if (membership?.archived) {
+            throw new TRPCError({ code: 'BAD_REQUEST', message: tr('Cannot edit archived membership') });
+        }
         return prisma.membership.delete({ where: { userId_groupId: data } });
     },
 
@@ -52,6 +62,7 @@ export const userMethods = {
             where: { id },
             include: {
                 memberships: {
+                    where: { archived: false },
                     include: { group: true, user: true, roles: true },
                     orderBy: { group: { name: 'asc' } },
                 },
@@ -69,7 +80,10 @@ export const userMethods = {
     },
 
     getMemberships: (id: string): Promise<MembershipInfo[]> => {
-        return prisma.membership.findMany({ where: { userId: id }, include: { group: true, user: true, roles: true } });
+        return prisma.membership.findMany({
+            where: { userId: id, archived: false },
+            include: { group: true, user: true, roles: true },
+        });
     },
 
     getGroupMembers: (groupId: string): Promise<User[]> => {
